separates base Scraper functionality into a new parent class
